Moved components into their own folders.
