+++ src/Components/ListItem/index.js
@@ -1,0 +1,10 @@
+function ListItem({ id, text, handleDelete }) {
+  return (
+    <li>
+      {text}
+      <button onClick={() => handleDelete(id)}>Delete</button>
+    </li>
+  );
+}
+
+export default ListItem;
